Move the template function to another file
